Admin can now edit and delete volunteer + minor fixes

+++ events.php
@@ -57,7 +57,7 @@ function monthRow($row) {
 	$year = $row['year'];
 	$month = $row['month'];
 	$nvolontari=$row['maxVolunteerNumber'];
-	$actions = "<a href='delete_month.php?year=$year?month=$month'>cancella</a>";
+	$actions = "<a onclick=\"return confirm('Sei sicuro di voler eliminare il mese selezionato? Tutte le prenotazioni e gli eventi associati verranno cancellati')\" href='delete_month.php?year=$year&month=$month'>cancella</a>";
 
 	$row = <<<EOF
 		<tr>
